Improve transforms test codebase - refactored compareTensorToPIL, _create_data, approxEqualTensorToPIL methods
