+++ test/test_functional_tensor.py
@@ -1,12 +1,9 @@
 import unittest
-import random
 import colorsys
 import math
 
-from PIL import Image
-from PIL.Image import NEAREST, BILINEAR, BICUBIC
-
 import numpy as np
+from PIL.Image import NEAREST, BILINEAR, BICUBIC
 
 import torch
 import torchvision.transforms as transforms
@@ -14,27 +11,10 @@ import torchvision.transforms.functional_tensor as F_t
 import torchvision.transforms.functional_pil as F_pil
 import torchvision.transforms.functional as F
 
+from common_utils import TransformsTester
 
-class Tester(unittest.TestCase):
 
-    def _create_data(self, height=3, width=3, channels=3, device="cpu"):
-        tensor = torch.randint(0, 255, (channels, height, width), dtype=torch.uint8, device=device)
-        pil_img = Image.fromarray(tensor.permute(1, 2, 0).contiguous().cpu().numpy())
-        return tensor, pil_img
-
-    def compareTensorToPIL(self, tensor, pil_image, msg=None):
-        pil_tensor = torch.as_tensor(np.array(pil_image).transpose((2, 0, 1)))
-        if msg is None:
-            msg = "tensor:\n{} \ndid not equal PIL tensor:\n{}".format(tensor, pil_tensor)
-        self.assertTrue(tensor.cpu().equal(pil_tensor), msg)
-
-    def approxEqualTensorToPIL(self, tensor, pil_image, tol=1e-5, msg=None):
-        pil_tensor = torch.as_tensor(np.array(pil_image).transpose((2, 0, 1))).to(tensor)
-        mae = torch.abs(tensor - pil_tensor).mean().item()
-        self.assertTrue(
-            mae < tol,
-            msg="{}: mae={}, tol={}: \n{}\nvs\n{}".format(msg, mae, tol, tensor[0, :10, :10], pil_tensor[0, :10, :10])
-        )
+class Tester(TransformsTester):
 
     def _test_vflip(self, device):
         script_vflip = torch.jit.script(F_t.vflip)
